Moved: PlayerColor to separate file
